refactor: API 타입 구조를 OpenAPI 스키마 중심으로 단순화
