+++ src/types/api.ts
@@ -1,4 +1,4 @@
-import type { components, paths } from './generated/schema';
+import type { components, paths } from './schema';
 
 export type ApiSchemas = components['schemas'];
 export type ApiPaths = paths;
