--- src/types/community.ts
@@ -1,5 +1,0 @@
-import type { ApiSchemas } from './api';
-
-export type CommunityPostRequest = ApiSchemas['CommunityPostRequestDTO'];
-export type CommunityCommentRequest =
-  ApiSchemas['CommunityPostCommentRequestDTO'];
